src: pcm: add pcm_fswitch plugin
This plugin makes it possible to open a different slave based on a
format. Further first device in the list is opened per default, then in
hw_params if a slave is found with the same format as requested the
previously opened slave is closed and the desired slave is opened.

With this plugin it is possible to only have a single device from the
application point of view, but still be able to redirect the audio based
on the format.

Further documentation on usage and configuration is inside the code.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>